+++ example_auto_nodes/__init__.py
@@ -1,4 +1,4 @@
 from .node_base.auto_node import AutoNode
 from .node_base.module_node import ModuleNode
-from .node_base.subgraph_node import SubGraphNode
+from .node_base.subgraph_node import SubGraphNode, RootNode
 from .subgraph_nodes import Publish
